Added script for doritos, merged spawningscripts into one, big, objectspawning script.
