--- Assets/Scripts/CloneMtDew.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4ab8391f45fe5cc449965f0bf34e51e0
-timeCreated: 1463997920
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
