+++ Assets/Scripts/CloneObjects.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 6338cb3cf6dbed84592bbbea0beee114
-timeCreated: 1464004648
+guid: 4550107064cd59948ad3c2a718c83f70
+timeCreated: 1464079862
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
